+++ test/crossfile/definition.lua
@@ -891,25 +891,69 @@ print(t.<?x?>)
     }
 }
 
---config.set('Lua.runtime.path', {
---    '/home/?.lua'
---})
---TEST {
---    {
---        path = '/home/a.lua',
---        content = [[
---return {
---    <!x!> = 1,
---}
---]],
---    },
---    {
---        path = 'b.lua',
---        content = [[
---local t = require 'a'
---print(t.<?x?>)
---        ]]
---    }
---}
+config.set('Lua.runtime.path', {
+    '/home/?.lua'
+})
+TEST {
+    {
+        path = '/home/a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
 
+config.set('Lua.runtime.pathStrict', true)
+config.set('Lua.runtime.path', {
+    './?.lua'
+})
+TEST {
+    {
+        path = 'a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.path', {
+    '/home/?.lua'
+})
+TEST {
+    {
+        path = '/home/a.lua',
+        content = [[
+return {
+    <!x!> = 1,
+}
+]],
+    },
+    {
+        path = 'b.lua',
+        content = [[
+local t = require 'a'
+print(t.<?x?>)
+        ]]
+    }
+}
+
+config.set('Lua.runtime.pathStrict', false)
 config.set('Lua.runtime.path', originRuntimePath)