+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 2.5.4
 * `FIX` [#848](https://github.com/sumneko/lua-language-server/issues/848)
+* `FIX` completion: incorrect cache
 
 ## 2.5.3
 `2021-12-6`