+++ changelog.md
@@ -3,6 +3,7 @@
 ## 2.5.4
 * `FIX` [#848](https://github.com/sumneko/lua-language-server/issues/848)
 * `FIX` completion: incorrect cache
+* `FIX` hover: always view string
 
 ## 2.5.3
 `2021-12-6`